🎨 Do not include detailed errors in the "stack" property of the error object
When the Metalsmith CLI fails, the "message" property value and "stack" property value of the thrown error object are displayed on the console.
In the case of a long error message, if the same information is duplicated, it will be difficult to read and debugging will be difficult.
To avoid this, include detailed errors only in the value of the message property.